requirements.txt
spotipy
yt-dlp
ytmusicapi
rich
python-dotenv

